refactor(service): update package infra
